Replaced math double functions with float functions

+++ colors/colors_operations2.c
@@ -6,7 +6,7 @@
 /*   By: amsaleh <amsaleh@student.42amman.com>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/19 17:26:35 by amsaleh           #+#    #+#             */
-/*   Updated: 2025/04/19 17:31:19 by amsaleh          ###   ########.fr       */
+/*   Updated: 2025/04/21 17:45:23 by amsaleh          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,8 +16,8 @@ t_colors	colorpow_f(t_colors *a, float val)
 {
 	t_colors	res;
 
-	res.red = pow(a->red, val);
-	res.green = pow(a->green, val);
-	res.blue = pow(a->blue, val);
+	res.red = powf(a->red, val);
+	res.green = powf(a->green, val);
+	res.blue = powf(a->blue, val);
 	return (res);
 }
